\name{eutil_error}
\alias{eutil_error}
\title{Generator object for the \code{\linkS4class{eutil_error}} class}
\usage{
  eutil_error(...)
}
\description{
  Generator object for the \code{\linkS4class{eutil_error}}
  class
}
\section{Methods}{
  \describe{ \item{\code{#all_empty()}:}{Are all error
  fields \code{NULL}} \item{\code{#check_errors(.Object,
  verbose = TRUE)}:}{check a \code{linkS4class{eutil}}
  object for errors} }
}
\keyword{classes}
\keyword{internal}

