% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efetch.R
\docType{methods}
\name{[,efetch,character,missing-method}
\alias{[,efetch,character,missing-method}
\alias{[[,efetch,character-method}
\title{EFetch accessors}
\usage{
\S4method{[}{efetch,character,missing}(x, i, j)

\S4method{[[}{efetch,character}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{efetch}} object containing XML data.}

\item{i}{An XPath expression specifying the XML nodes to extract.}

\item{j}{Ignored.}
}
\value{
An XML node set.
}
\description{
Extract XML nodes from an \code{\linkS4class{efetch}} object.
}
\examples{
\dontrun{
p <- efetch("195055", "protein", "gp", "xml")
p['//GBFeature[GBFeature_key="mat_peptide"]//GBQualifier_value']
}
}
