% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einfo.R
\name{[,einfo,ANY,missing-method}
\alias{[,einfo,ANY,missing-method}
\alias{[[,einfo,ANY-method}
\title{EInfo accessors}
\usage{
\S4method{[}{einfo,ANY,missing}(x, i, j)

\S4method{[[}{einfo,ANY}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{einfo}} object.}

\item{i}{Numeric or character indices specifying the elements to extract.}

\item{j}{Ignored.}
}
\value{
A list.
}
\description{
Extract parts of a parsed \code{\linkS4class{einfo}} object.
}
\examples{
\dontrun{
e <- einfo("pubmed")
e[1:5]
e["Description"]
e[["Links"]]

e2 <- einfo("pubmed", retmode = 'json')
e2[["header"]]
e2[["einforesult"]][["dbinfo"]][["description"]]
}
}
\seealso{
\code{\link[base]{Extract}}
}
