\name{ncbi_retrieval_type}
\alias{ncbi_retrieval_type}
\title{Set the NCBI rettype}
\usage{
  ncbi_retrieval_type(db, rettype = NULL, retmode = NULL)
}
\arguments{
  \item{db}{A valid NCBI database.}

  \item{rettype}{Optional.}

  \item{retmode}{Optional.}
}
\description{
  Set the NCBI rettype
}
\keyword{internal}

