\docType{methods}
\name{performQuery}
\alias{performQuery}
\alias{performQuery,eutil-method}
\title{performQuery}
\arguments{
  \item{x}{An \code{\linkS4class{eutil}} object.}

  \item{method}{One of \dQuote{GET} or \dQuote{POST}.}

  \item{...}{Further arguments passed on to methods.}
}
\description{
  performQuery
}
\keyword{internal}

