% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epost.R, R/esearch.R, R/eutil.R
\docType{methods}
\name{querykey,epost-method}
\alias{querykey}
\alias{querykey,entrez_uid-method}
\alias{querykey,epost-method}
\alias{querykey,esearch-method}
\title{querykey}
\usage{
\S4method{querykey}{epost}(x, ...)

\S4method{querykey}{entrez_uid}(x, ...)

\S4method{querykey}{esearch}(x, ...)

querykey(x, ...)
}
\arguments{
\item{x}{An \code{\linkS4class{eutil}} object.}

\item{...}{Further arguments passed on to methods.}
}
\value{
An integer or \code{NA}.
}
\description{
An integer query key returned by an ESearch, EPost or ELink call if
the History server was used. Otherwise \code{NA}.
}
\section{Methods (by class)}{
\itemize{
\item \code{epost}: 

\item \code{entrez_uid}: 

\item \code{esearch}: 
}}
\examples{
\dontrun{
e <- esearch("Mus musculus", "taxonomy", usehistory = TRUE)
querykey(e)
}
}
\seealso{
\code{\link{einfo}}, \code{\link{esearch}}, \code{\link{esummary}},
   \code{\link{efetch}}, \code{\link{elink}}, \code{\link{epost}},
   \code{\link{egquery}}, \code{\link{espell}}, \code{\link{ecitmatch}}.
}

