% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elink.R, R/esearch.R, R/eutil.R
\docType{methods}
\name{uid,entrez_linkset-method}
\alias{uid}
\alias{uid,elink-method}
\alias{uid,entrez_linkset-method}
\alias{uid,entrez_uid-method}
\alias{uid,esearch-method}
\title{uid}
\usage{
\S4method{uid}{entrez_linkset}(x, ...)

\S4method{uid}{elink}(x, ...)

\S4method{uid}{entrez_uid}(x, ...)

\S4method{uid}{esearch}(x, ...)

uid(x, ...)
}
\arguments{
\item{x}{An \code{\linkS4class{eutil}} object.}

\item{...}{Further arguments passed on to methods.}
}
\value{
A character vector.
}
\description{
Retrieve the list of UIDs returned by a call to ESearch or ELink.
}
\section{Methods (by class)}{
\itemize{
\item \code{entrez_linkset}: 

\item \code{elink}: 

\item \code{entrez_uid}: 

\item \code{esearch}: 
}}
\examples{
\dontrun{
e <- esearch("Mus musculus", "taxonomy")
uid(e)
}
}
\seealso{
\code{\link{esearch}}, \code{\link{elink}}.
}

