% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{xvalue}
\alias{xvalue}
\title{Extract the content of XML leaf nodes}
\usage{
xvalue(doc, path, as = "character", default = NA_character_, ...)
}
\arguments{
\item{doc}{An object of class \code{XMLInternalDocument}.}

\item{path}{An XPath expression.}

\item{as}{Mode of return value (\code{character}, \code{integer}, \code{numeric}).}

\item{default}{Default return value.}

\item{...}{Arguments passed to \code{\link[XML]{xpathApply}}.}
}
\description{
Extract the content of XML leaf nodes
}
\keyword{internal}

