% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{database,entrez_linkset-method}
\alias{database}
\alias{database,entrez_linkset-method}
\alias{database,entrez_uid-method}
\alias{database,eutil-method}
\title{database}
\usage{
\S4method{database}{entrez_linkset}(x, ...)

\S4method{database}{entrez_uid}(x, ...)

database(x, ...)

\S4method{database}{eutil}(x, ...)
}
\arguments{
  \item{x}{An \code{\linkS4class{eutil}} object.}

  \item{...}{Further arguments passed on to methods.}
}
\value{
A character string.
}
\description{
Retrieve the target database name from an \code{\linkS4class{eutil}} object.
}
\examples{
\dontrun{
e <- esearch("Mus musculus", "taxonomy")
database(e)
}
}
\seealso{
\code{\link{einfo}}, \code{\link{esearch}}, \code{\link{esummary}},
   \code{\link{efetch}}, \code{\link{elink}}, \code{\link{epost}},
   \code{\link{egquery}}, \code{\link{espell}}, \code{\link{ecitmatch}}.
}

