\name{egquery}
\alias{egquery}
\title{egquery}
\usage{
  egquery(term)
}
\arguments{
  \item{term}{A valid Entrez text query.}
}
\value{
  An \code{\linkS4class{egquery}} object.
}
\description{
  \code{egquery} retrieves the number of records in all
  Entrez databases for a single text query
}
\details{
  See the official online documentation for NCBI's
  \href{http://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.EGQuery}{EUtilities}
  for additional information.
}
\examples{
###
}

