% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{[,einfo,ANY-method}
\alias{[,einfo,ANY-method}
\alias{[[,einfo,ANY-method}
\title{EInfo accessors}
\usage{
\S4method{[}{einfo,ANY}(x, i)

\S4method{[[}{einfo,ANY}(x, i)
}
\arguments{
  \item{x}{An \code{\linkS4class{einfo}} object.}

  \item{i}{Numeric or character indices specifying the
  elements to extract.}
}
\value{
A list.
}
\description{
Extract parts of a parsed \code{\linkS4class{einfo}} object.
}
\examples{
e <- einfo("pubmed")
e[1:5]
e["Description"]
e[["Links"]]
}
\seealso{
\code{\link[base]{Extract}}
}

