% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/einfo.R
\name{einfo}
\alias{einfo}
\title{einfo - getting database statistics and search fields}
\usage{
einfo(db = NULL, version = NULL, retmode = "xml")
}
\arguments{
\item{db}{A valid NCBI database name. If \code{NULL}, a list of all current NCBI
databases is returned.}

\item{version}{Specifies version 2.0 EInfo XML. The only supported value is '2.0'.}

\item{retmode}{'xml' (default) or 'json'.}
}
\value{
An \code{\linkS4class{einfo}} object.
}
\description{
\code{einfo} performs calls to the NCBI EInfo utility to retrieve the names
of all valid Entrez databases, or, if \code{db} is provided,
to retrieve statistics for a single database, including lists of indexing
fields and available link names.
}
\details{
See the official online documentation for NCBI's
\href{http://www.ncbi.nlm.nih.gov/books/NBK25499/\#chapter4.EInfo}{EUtilities}
for additional information.
}
\examples{
## Fetch a list of all current Entrez database names
einfo()

## Fetch statistics for an Entrez database and parse
## the data into a data.frame
x <- einfo("gene")
content(x, as = "parsed")

## Fetch statistics for an Entrez database in JSON format
## and parse the data into a list
x <- einfo("pubmed", retmode = "json")
content(x, as = "parsed")
}
\seealso{
\code{\link{content}}, \code{\link{getUrl}}, \code{\link{getError}}.
}

