% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elink.R
\docType{methods}
\name{[,elink,numeric-method}
\alias{[,elink,character-method}
\alias{[,elink,numeric-method}
\title{ELink Accessors}
\usage{
\S4method{[}{elink,numeric}(x, i)

\S4method{[}{elink,character}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{elink}} object.}

\item{i}{Numeric or character indices.}
}
\value{
A \code{\linkS4class{entrez_linkset}} object.
}
\description{
Extract UIDs from an \code{\link{elink}} object.
}
\examples{
\dontrun{
e <- elink(c("34577062", "24475906"), dbFrom = "nuccore")
e[1]
}
}

