% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/esearch.R
\docType{methods}
\name{[,esearch,numeric-method}
\alias{[,esearch,numeric-method}
\title{ESearch Accessors}
\usage{
\S4method{[}{esearch,numeric}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{esearch}} object.}

\item{i}{Numeric indices.}
}
\value{
A \code{\linkS4class{entrez_uid}} object.
}
\description{
Extract UIDs from an \code{\link{esearch}} object.
}
\examples{
\dontrun{
e <- esearch("Mus musculus", "protein", retmax = 20)
e[1:5]
## pass the subset directly on to esummary or efetch
content(esummary(e[1:5]), "parsed")
}
}

