% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{espell}
\alias{espell}
\title{espell - retrieving spelling suggestions}
\usage{
espell(term, db = "nuccore")
}
\arguments{
\item{term}{An Entrez text query.}

\item{db}{An Entrez database.}
}
\value{
An \code{\linkS4class{espell}} object.
}
\description{
For a text query retrieve an XML containing the original query
and spelling suggestions.
}
\examples{
###
}

