% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/esummary.R
\docType{methods}
\name{[,esummary,character-method}
\alias{[,esummary,character-method}
\alias{[[,esummary,character-method}
\title{ESummary accessors}
\usage{
\S4method{[}{esummary,character}(x, i)

\S4method{[[}{esummary,character}(x, i)
}
\arguments{
\item{x}{An \code{\linkS4class{esummary}} object.}

\item{i}{An XPath expression.}
}
\value{
An XML node set.
}
\description{
Extract XML nodes from an \code{\linkS4class{esummary}} object.
}
\examples{
\dontrun{
ds <- esummary("470338", "protein")
ds["//Slen/node()"]

library("XML")
as.numeric(xmlValue(ds[["//Slen"]]))
}
}

