\name{eutil_error-class}
\alias{eutil_error-class}
\title{Class \code{"eutil_error"}}
\description{
  A container for handling errors when trying to parse XML
  files returned by Entrez.
}
\section{Fields}{
  \describe{ \item{\code{error}:}{} \item{\code{errmsg}:}{}
  \item{\code{wrnmsg}:}{} }
}

\section{Extends}{
  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
###
}
\seealso{
  \code{\link{eutil_error}}, \code{\linkS4class{eutil}}.
}
\keyword{classes}

