% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ncbi_retrieval_type}
\alias{ncbi_retrieval_type}
\title{Set the NCBI rettype}
\usage{
ncbi_retrieval_type(db, rettype = NULL, retmode = NULL)
}
\arguments{
\item{db}{A valid NCBI database.}

\item{rettype}{Optional.}

\item{retmode}{Optional.}
}
\description{
Set the NCBI rettype
}
\keyword{internal}

