\name{performQuery}
\alias{performQuery}
\alias{performQuery,eutil-method}
\alias{performQuery-methods}
\title{performQuery}
\usage{
  \S4method{performQuery}{eutil}(x, method = "GET", ...)
}
\arguments{
  \item{x}{An \code{\linkS4class{eutil}} object.}

  \item{method}{One of \dQuote{GET} or \dQuote{POST}.}

  \item{...}{Further arguments passed on to methods.}
}
\description{
  performQuery
}
\keyword{internal}

