% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutil.R
\docType{methods}
\name{performQuery}
\alias{performQuery}
\alias{performQuery,eutil-method}
\title{performQuery}
\usage{
performQuery(x, method = "GET", ...)

\S4method{performQuery}{eutil}(x, method = "GET", ...)
}
\arguments{
\item{x}{An \code{\linkS4class{eutil}} object.}

\item{method}{One of \dQuote{GET} or \dQuote{POST}.}

\item{...}{Further arguments passed on to methods.}
}
\description{
performQuery
}
\section{Methods (by class)}{
\itemize{
\item \code{eutil}: Perform an Entrez query using either http GET or
POST requests.
}}

\keyword{internal}
