% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{webenv,epost-method}
\alias{webenv}
\alias{webenv,entrez_uid-method}
\alias{webenv,epost-method}
\alias{webenv,esearch-method}
\title{webenv}
\usage{
\S4method{webenv}{epost}(x, ...)

\S4method{webenv}{entrez_uid}(x, ...)

\S4method{webenv}{esearch}(x, ...)

webenv(x, ...)
}
\arguments{
\item{x}{An \code{\linkS4class{eutil}} object.}

\item{...}{Further arguments passed on to methods.}
}
\value{
A character string or \code{NA}.
}
\description{
Retrieve the Web environment string returned from an ESearch, EPost or ELink call.
\code{NA} if the History server was not used.
}
\examples{
\dontrun{
e <- esearch("Mus musculus", "taxonomy", usehistory = TRUE)
webenv(e)
}
}
\seealso{
\code{\link{einfo}}, \code{\link{esearch}}, \code{\link{esummary}},
   \code{\link{efetch}}, \code{\link{elink}}, \code{\link{epost}},
   \code{\link{egquery}}, \code{\link{espell}}, \code{\link{ecitmatch}}.
}

